fix dtype

+++ neural_speed/core/layers/ne_bestla.cpp
@@ -78,7 +78,7 @@ void bestla_packweight_copyattr(const float* f32ptr, void* dstptr, int n, int k,
   safe_delete(wtmp);
 }
 
-void bestla_layernormalization(size_t norm_count, size_t norm_size, bool isrms, float epsilon, const float* FpIn,
+void bestla_layernormalization(int norm_count, int norm_size, bool isrms, float epsilon, const float* FpIn,
                                float* FpOut) {
   BTLALayerNorm(norm_count, norm_size, isrms, epsilon, FpIn, FpOut, ne_threading::get());
 }
